Coerce set theory filters to always return lists, not sets.
The set theory filters, used with expressions like:

  {{ list1 | unique }}
  {{ list1 | union(list2) }}

… were introduced in Ansible 1.4, and are implemented using Python's set
operators, and return Python set objects.

When these set objects are serialised as strings and returned, they are
rendered using Python's default syntax:

  set(['a','b','c'])

… which can be interpreted properly by Ansible, e.g. when used as a
parameter in a with_items clause.

However, if you attempt to use these as an input to a second pass of the
set theory filters, then unexpected behaviour occurs.

For example, if you were to evaluate (either indirectly, or via variable
substitution) the expression:

  {{ set(['a','b','c']) | unique }}

… it would not produce the result:

  set(['a','b','c'])

… as you might expect, but would instead produce the result:

  set(['a', 'c', 'b', 'e', "'", ')', '(', ',', 's', 't', '[', ']'])

Which violates the Principle of Least Surprise.

To avoid this, this patch modifies each of the set operators added in
Ansible 1.4 to coerce their return types from <set> to <list>, so that
when serialised they produce the string:

  ['a','b','c']

… rather than the equivilent set notation.  Jinja2 knows how to interpret
list literals correctly, so will interpret the above elements in the
manner that the end-user would normally expect.

Signed-off-by: David McBride <dwm37@cam.ac.uk>