Add 'interpolate' option to hash_behaviour.
This commit extends the hash_behaviour configuration option, with a new
'interpolate' mode, to complement the existing 'merge' and 'replace'
modes.

The hash_behaviour becomes significant when a variable defined in one
place overrides some existing definition.  By default, the 'replace'
logic is used: the previous definition is lost, and replaced with
the new.

The other previous option was to use 'merge', which would unify the set
of keys in dictionaries where possible.

The new 'interpolate' mode allows a user to express, on a case by case
basis, how they want variables to be redefined — by allowing child
definitions to be expressed in terms of their parent using Jinja2
expressions.

Previously, all Jinja2 evaluation was delegated until the point of use:
it was only when a variable containing a Jinja2 expression was referenced
in a playbook that Ansible would cause it to be resolved.

With this mode, however, each individual var file is evaluated at
load-time, and the previous state of any variables defined at that point
is made available under the 'parent' var.

For example, you could define a host_group var foo:

  ---
  # group_vars/group
  foo: ['1','2','3']

... and then augment this value in a host-specific var file:

  ---
  # host_vars/member_of_group
  foo: "{{ parent['foo'] | union( ['4','5','6'] )}}"

In this example, foo would evaluate to the list ['1','2','3','4','5','6']
for the host "member_of_group", but just ['1','2','3'] for all other
members of "group".

Signed-off-by: David McBride <dwm37@cam.ac.uk>